refactor(daemon): keep the workspace-summary validator inside the pinned relation-query validator module
