+++ packages/daemon/src/protocol/daemon-protocol.contract.ts
@@ -924,4 +924,4 @@ export {
   validateDaemonTaskDocumentList,
 } from "./daemon-protocol-validate-results.ts";
 export { isolateDaemonTaskSnapshotRows, validateDaemonTaskSnapshotList } from "./daemon-protocol-validate-task.ts";
-export { validateDaemonWorkspaceSummary } from "./daemon-protocol-validate-workspace.ts";
+export { validateDaemonWorkspaceSummary } from "./daemon-protocol-validate-relation-query.ts";
